requirements.txt
altair==5.5.0
arabic-reshaper==3.0.0
attrs==24.3.0
blinker==1.9.0
Brotli==1.1.0
cachetools==5.5.0
certifi==2024.12.14
cffi==1.17.1
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
cssselect2==0.7.0
fonttools==4.55.3
gitdb==4.0.12
GitPython==3.1.44
idna==3.10
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
narwhals==1.20.1
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
protobuf==5.29.2
pyarrow==18.1.0
pycparser==2.22
pydeck==0.9.1
pydyf==0.11.0
Pygments==2.18.0
pyphen==0.17.0
pytesseract==0.3.13
python-bidi==0.6.3
python-dateutil==2.9.0.post0
pytz==2024.2
referencing==0.35.1
requests==2.32.3
rich==13.9.4
rpds-py==0.22.3
six==1.17.0
smmap==5.0.2
streamlit==1.41.1
tenacity==9.0.0
tinycss2==1.4.0
tinyhtml5==2.0.0
toml==0.10.2
tornado==6.4.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
weasyprint==63.1
webencodings==0.5.1
zopfli==0.2.3.post1
psycopg2==2.9.3
python-dotenv==0.19.1
openai==1.59.7
openpyxl==3.1.5
